+++ software/src/iec61851.h
@@ -47,6 +47,6 @@ void iec61851_tick(void);
 
 uint32_t iec61851_get_ma_from_pp_resistance(void);
 uint32_t iec61851_get_max_ma(void);
-uint16_t iec61851_get_duty_cycle_for_ma(uint32_t ma);
+float iec61851_get_duty_cycle_for_ma(uint32_t ma);
 
 #endif
